+++ packages/sandbox-e2b/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @sandagent/sandbox-e2b
 
+## 0.9.16-beta.5
+
+### Patch Changes
+
+- @sandagent/manager@0.9.16-beta.5
+
+## 0.9.16-beta.4
+
+### Patch Changes
+
+- @sandagent/manager@0.9.16-beta.4
+
+## 0.9.16-beta.3
+
+### Patch Changes
+
+- @sandagent/manager@0.9.16-beta.3
+
+## 0.9.16-beta.2
+
+### Patch Changes
+
+- @sandagent/manager@0.9.16-beta.2
+
+## 0.9.16-beta.1
+
+### Patch Changes
+
+- @sandagent/manager@0.9.16-beta.1
+
+## 0.9.16-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @sandagent/manager@0.9.16-beta.0
+
 ## 0.9.12
 
 ### Patch Changes